dump_smp_affinity: don't log data if the fread failed, make ret size_t too
The fread call may potentially fail and return 0 to signify end of file,
an error or zero data read. When there is a potential fread failure don't
log the empty/uninitialized string str.  Make variable ret size_t to match
the return type of fread to ensure the code is type safe.

Signed-off-by: Colin Ian King <colin.i.king@gmail.com>