Integrate workload type
Integrate workload type to prevent polling in idle and also exit
for non idle.

We can't blindly use workload type to enter exit LPM. Ideally
when workload type is sustained or bursty, we can just exit LPM
else enter LPM.

But the workload type is battery life on busy LPE core CPUs, whereas
on non LPE or p-cores it will be sustained or bursty.

So idea is to rely on current utilization method to enter/exit when
polling is enabled. When polling is not enabled that means system was
idle, then simply exit on non idle workload type and enable polling.

When exit in AUTO mode, existing utilization based algorithm takes
over.