Process perf=eff=0 for non surivability mode
When there are CPUs in HFI table then the current logic only
adds CPUs with EFF=1024 to the cpumask for HFI. This will exclude
CPUs with PERF=EFF=0. This means already accounting for this case.

But when the HFI table only contains PERF=EFF=0 without any
EFF=255 CPUs, then also exclude those CPUs from cpu mask. When this
happens copy the banned cpumask to HFI cpumask, this way even
PERF=EFF=0 is treated in the same way.

Originally-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>