Build default config states
System are in two different states when in and out of Low Power Mode.
Each of the states has its own requirement of EPP/EPB/ITMT setting and
IRQ_affinity/Task_affinity.

When utilization monitor is enabled and there is no config states
described in the config file, convert the above in-LPM and out-of-LPM
states as two default config states. So that we can share the same code
no matter if config states are available in the config file or not.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>