Implement util based config state support
Implement util based config state support.

When utilization monitor is activated, the deepest state that meets
the current system/cpu utilization will be choosed.

When utilization monitor is enabled and there is no config states
described in the config file, convert the in-LPM and out-of-LPM states
as two default config states.

Originally-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>
Signed-off-by: Zhang Rui <rui.zhang@intel.com>