Restore ITMT on USER_EXIT
There are multiple issues in restoring:
- The init_itmt() is not called hence ITMT flag is not stored
- On USER_EXIT to restore depends on lpmd_config.ignore_itmt
Even if the global lpmd_config.ignore_itmt is not set, individual
state may have set the flag. For example if there is a state
in which ITMT flag is set to 0, and USER_EXIT happens in this
state then flag will be 0 even after exit.

Hence call init_itmt() at the power up and on USER_EXIt always call
with argument SETTING_RESTORE, similar to what set_lpm_irq() is
doing here.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>