Introduce parameter "--ignore-platform-check"
current code uses a strict platform check to launch the service on
certain platforms that the ILEO has been verified to work.
This includes,
1. the platform must be a hybrid platform (via CPUID)
2. the platform must be a mobile platform (via ACPI)
3. the platform must be in an allowed list (via CPU model)

But this also blocks other users who want to experience it on the
platforms that don't meet all these criteria.

Introduce parameter "--ignore-platform-check" to ignore such checks.

Signed-off-by: Zhang Rui <rui.zhang@intel.com>