Add Power Slider support
Refer to the following link about the power slider support:
https://lore.kernel.org/linux-pm/20250813185530.635096-2-srinivas.pandruvada@linux.intel.com/

Add necessary support in the LPMD to set power slider balance and
offset value. These values can be defined in the platform specific config file
for both AC and DC. These knobs can be also specific to a state in the
config file.

These knobs are:
BalanceSliderAC, BalanceSliderDC, SliderOffsetAC, SliderOffsetDC

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>