Workload type notification without proxy
There is a bug in setting poll_fds[] for workload type file descriptor.
First of idx_wlt_fd is not assigned to the correct file descriptor and also
workload type poll() requires POLLPRI.

Without this if workload type without proxy is enabled it will exit
poll() and spin in the loop.

Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>