lpmd_config: Rename island number cores tags
Redo these config options into a number of active cores of a given type
to specify per state. This is a preparation for a configuration feature
to name a number of a specific core type instead of hardcoding the
actual core indices.

Remove the excess island number cores tags since they're not used
anywhere.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>