lpmd_state_machine: Count system of specifing active cores
Instead of supplying the specific cpu core indices in the ActiveCpus
tags in the configuration it would be more convenient to only put down
the desired number of cores of a specific type / types.

Add a simple system of core type bitmasks from which the program can
later derive cpumasks based on the usage of ActiveP/E/Lcores in the
configuration states.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>