lpmd_config: Skip platform check if a matching config file exists
A platform that's not added to the model list in the source code is
discarded early when launching the program.

For easier scaling it's beneficial to accept platforms that have a
matching configuration file in the appropriate directory.

Skip matching the platform model to the source code defined list of
supported models if a matching config file exists.

Signed-off-by: Maciej Wieczor-Retman <maciej.wieczor-retman@intel.com>
[srinivas: Fix coverity error for uninitialized file_name]
Signed-off-by: Srinivas Pandruvada <srinivas.pandruvada@linux.intel.com>